fix: make the build work on Linux and Windows, not just macOS
Everything in this branch had only ever been built and tested on macOS. The
first CI run failed on both other platforms, in eight distinct ways, none of
which a macOS build can produce. All of them are now reproduced and verified
locally in a Linux container; the two Windows ones are diagnosed from the CI
logs and fixed at their root cause.

Linux, all reproduced under Ubuntu 24.04 with GCC:

- `nanosleep` and `getpid` are POSIX, and glibc hides both under a strict
  -std=c11 unless a feature-test macro asks for them. Apple's headers expose
  them regardless, so src/http/client.c did not compile on Linux at all.
- GCC does not accept a `(void)` cast as suppression for a
  warn_unused_result function. Six `tamga_uuid_normalize` calls are now
  checked rather than discarded, which is better than suppressing: the
  comments claimed the calls could not fail, and an enforced invariant
  survives the earlier validation being moved. The one genuinely deliberate
  discard, in transport_curl.c, is assigned to a variable instead.
- Three -Wconversion/-Wsign-conversion errors in gcm.c, ed25519.c and bn.c,
  visible only in the ASan configuration because its optimisation level
  changes what GCC's range analysis can prove.
- tests/c/abi_surface_test.c stored function pointers in a void * array. ISO C
  forbids that conversion and GCC rejects it under -Wpedantic — so the file
  whose whole job is to be the portable ABI gate was the one that did not
  build portably. It uses a generic function-pointer type now.
- Scripts/check-coverage.sh had two bugs that cancelled out on macOS. Its tool
  detection was `command -v X || command -v xcrun && echo ...`, which parses
  as ((A || B) && C): where the tool IS on PATH, A prints the path and C
  appends a second line, so the command ran as `llvm-profdata xcrun
  llvm-profdata merge`. And llvm-cov cannot read a static archive on Linux at
  all; it reads the shared library instead, with the archive as the fallback.
- The clang-format/clang-tidy job installed into the system Python, which
  Debian and Ubuntu refuse under PEP 668. It uses a virtualenv now.
- clang-tidy's Annex K check (memset_s/memcpy_s) fires on Linux and not on
  macOS. Annex K is optional and neither glibc nor Apple's libc implements
  it, so taking its advice would not compile on either target; the check is
  disabled with that reason recorded.

Windows, from the CI logs:

- Both library targets were named `tamga`, and on Windows a DLL's import
  library is also a .lib — so the static archive overwrote the shared
  library's import library, and everything linking the DLL then linked the
  static archive and failed on unresolved __imp_ symbols. All four ABI tests
  died this way. The static library is named tamga_static on Windows now;
  elsewhere the extensions differ and there was never a collision.
- The whitebox test targets compile the library's sources directly rather
  than linking a library, so they never received TAMGA_STATIC_LIB and MSVC
  saw dllimport declarations next to their own definitions — 64 instances of
  C4273, fatal under /WX. They define it themselves now, as does the fuzz
  library, which has the same shape.

Verified: macOS, and under Ubuntu 24.04 the Debug, ASan+UBSan+LeakSanitizer,
TAMGA_HTTP=none (with libcurl not installed — ldd shows libc alone), coverage
and fuzz configurations, all 28 tests passing in each. Windows remains CI-only.

+++ src/crypto/ed25519.c
@@ -60,7 +60,9 @@ static void tamga_sc_reduce512(const unsigned char in[64], unsigned char out[32]
     tamga_sc_load_l(l);
 
     for (bit = 511; bit >= 0; bit--) {
-        uint32_t carry = (uint32_t)((in[bit / 8] >> (bit % 8)) & 1u);
+        /* Same widening as in gcm.c, and the shift count is made unsigned
+         * too since `bit` is a signed loop counter. */
+        uint32_t carry = ((uint32_t)in[bit / 8] >> (unsigned int)(bit % 8)) & 1u;
         uint32_t candidate[8];
         uint64_t borrow = 0u;
